third_party/WebKit/LayoutTests/inspector/initial-modules-load-expected.txt

This test validates initial set of loaded modules.


Running: testInitialLoad
Loaded modules:
    bindings
    common
    components
    dom_extension
    emulation
    extensions
    help
    host
    main
    network_conditions
    object_ui
    persistence
    platform
    protocol
    sdk
    services
    ui
    workspace

Running: testCreateElementsPanel
Loaded modules:
    bindings
    color_picker
    common
    components
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    help
    host
    inline_editor
    main
    network_conditions
    object_ui
    persistence
    platform
    protocol
    sdk
    services
    ui
    workspace

Running: testCreateNetworkPanel
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    help
    host
    inline_editor
    main
    network
    network_conditions
    object_ui
    perf_ui
    persistence
    platform
    protocol
    sdk
    services
    source_frame
    text_editor
    ui
    workspace

Running: testShowSourcesPanel
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    help
    host
    inline_editor
    main
    network
    network_conditions
    object_ui
    perf_ui
    persistence
    platform
    protocol
    quick_open
    sdk
    services
    snippets
    source_frame
    sources
    text_editor
    ui
    workspace

Running: testOpenUISourceCode
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    help
    host
    inline_editor
    main
    network
    network_conditions
    object_ui
    perf_ui
    persistence
    platform
    protocol
    quick_open
    sdk
    services
    snippets
    source_frame
    sources
    text_editor
    ui
    workspace
